feat: 약속 상세 seniorId 연결
