+++ src/pages/promiseList/components/RecentCard.tsx
@@ -13,11 +13,10 @@ interface RecentCardPropType {
   recentAppointment?: profileCardDataType;
   appointmentNum: number;
   nickname: string;
-  handleSetIsDetailClicked: (type: boolean, id: number) => void;
 }
 
 const RecentCard = (props: RecentCardPropType) => {
-  const { userRole, recentAppointment, appointmentNum, nickname, handleSetIsDetailClicked } = props;
+  const { userRole, recentAppointment, appointmentNum, nickname } = props;
   const { diffText, diff, dDayDiff } = useCountdown(recentAppointment?.date, recentAppointment?.startTime);
   const [isEnterBtnClicked, setIsEnterBtnClicked] = useState(false);
   const [, setGoogleMeetLink] = useState('');
@@ -47,8 +46,6 @@ const RecentCard = (props: RecentCardPropType) => {
             profileCardData={recentAppointment}
             isarrow="true"
             myNickname={nickname}
-            detail="detail"
-            handleSetIsDetailClicked={handleSetIsDetailClicked}
           />
           <PromiseTimerBtn
             isActive={diff <= 0}
